Skeleton C Program
The main will handle command line options and call the encoder
or decoder depending on how the program is invoked. The encoder
and decoder functions are stubbed out with argument checking
and not much else. The Makefile can build the project, clean up
object files, create a tar file source distribution and clobber
all executables, tar files and directories created for the source
distribution.

I was planning to call mmecode and mmdecode with a pointer to an
options_t struct, so options_t is defined in main.h to make that
definition available in mmencode.c and mmdecode.c, but I'm changing
my mind and will move option_t to main.c since it will only be used
there.